Rename ternary_state_controller->ternary_controller
