+++ ternary_controller/include/ternary_controller/ternary_state_interface.h
